+++ src/runtime/components/Layer.vue
@@ -1,15 +1,22 @@
 <!-- eslint-disable vue/multi-word-component-names -->
 <script setup lang="ts">
-import { AnyLayer, AnySourceData, Layer, MapMouseEvent } from "mapbox-gl";
-import { inject, onMounted } from "vue";
-import { useMapbox } from "../composables/useMapbox";
-import { computed, onUnmounted, watch, getCurrentInstance } from "#imports";
+import { AnyLayer, Layer, MapMouseEvent } from "mapbox-gl";
+import { inject } from "vue";
+import { computed, onUnmounted, watch, useMapboxInstance, useMapbox, useMapboxBeforeLoad, useAttrs } from "#imports";
+import { whenever } from "@vueuse/core";
 
 interface Props {
-    sourceId?: string;
-    source?: AnySourceData;
     layer: AnyLayer;
-    beforeLayer?: string
+    beforeLayer?: string;
+
+    onMousedown?: Function;
+    onMouseup?: Function;
+    onMouseover?: Function;
+    onMousemove?: Function;
+    onClick?: Function;
+    onDblclick?: Function;
+    onMouseenter?: Function;
+    onMouseleave?: Function;
 }
 const props = defineProps<Props>();
 
@@ -27,85 +34,51 @@ const emit = defineEmits<{
 const mapId = inject<string>("MapID");
 if (!mapId) throw "Mapbox Controls must be placed inside a Map component";
 
-if (props.source || props.sourceId) console.warn("source & sourceId props in MapboxLayer are deprecated");
+const attrs = useAttrs();
+if (attrs.source || attrs.sourceId) throw "source & sourceId props in MapboxLayer are deprecated";
 
-// Vue magic so we can see which events are being listened to
-// Mapbox slows down significantly with each event listener, even if we do nothing.
-// This may lead to bugs down the line with events getting added to the component after it's mounted.
-const vnodeProps = getCurrentInstance()?.vnode.props
+watch(() => {
+    return Object.entries(props).filter((arr) => arr[0].toString().startsWith('on'))
+}, (value) => {
+    for (const [key, v] of value) {
+        if (v) {
+            const eventName = key.toString().replace('on', '').toLowerCase();
+            useMapboxBeforeLoad(mapId, (map) => {
+                map.on(eventName as keyof typeof emit, props.layer.id, (e) => {
+                    emit(eventName as keyof typeof emit, e);
+                })
+            })
+        }
+    }
+}, {immediate: true});
 
-onMounted(() => {
-    useMapbox(mapId, (map) => {
-        const sourceExists = computed(() => {
-            return !!map?.getSource(
-                (props.layer as Layer).source?.toString() || props.sourceId || ""
-            );
-        });
-        function addLayer() {
-            if (!sourceExists.value) {
-                watch(sourceExists, addLayer);
-                return;
-            }
-            if (props.beforeLayer && map.getLayer(props.beforeLayer)) {
-                map?.addLayer(props.layer, props.beforeLayer);
-            } else {
-                if (props.beforeLayer) {
-                    watch(() => map.getLayer(props.beforeLayer || ''), () => {
-                        map?.moveLayer(props.layer.id, props.beforeLayer);
-                    })
-                }
-                map?.addLayer(props.layer);
-            }
-        }
-        function addSource() {
-            if (props.source && props.sourceId) {
-                map?.addSource(props.sourceId, props.source);
-            }
-        }
+const mapRef = useMapboxInstance(mapId);
+const sourceExists = computed(() => {
+    return !!mapRef.value?.getSource(
+        (props.layer as Layer).source?.toString() || ""
+    );
+});
 
-        addSource();
-        addLayer();
-        if (vnodeProps?.onMousedown) {
-            map.on("mousedown", props.layer.id, (e) => {
-                emit("mousedown", e);
-            });
+useMapbox(mapId, (map) => {
+    function addLayer() {
+        if (!sourceExists.value) {
+            // This causes layer to be reloaded whenever the source is.
+            whenever(sourceExists, addLayer);
+            return;
         }
-        if (vnodeProps?.onMouseup) {
-            map.on("mouseup", props.layer.id, (e) => {
-                emit("mouseup", e);
-            });
+        if (props.beforeLayer && map.getLayer(props.beforeLayer)) {
+            map?.addLayer(props.layer, props.beforeLayer);
+        } else {
+            if (props.beforeLayer) {
+                watch(() => map.getLayer(props.beforeLayer || ''), () => {
+                    map?.moveLayer(props.layer.id, props.beforeLayer);
+                })
+            }
+            map?.addLayer(props.layer);
         }
-        if (vnodeProps?.onMouseover) {
-            map.on("mouseover", props.layer.id, (e) => {
-                emit("mouseover", e);
-            });
-        }
-        if (vnodeProps?.onMousemove) {
-            map.on("mousemove", props.layer.id, (e) => {
-                emit("mousemove", e);
-            });
-        }
-        if (vnodeProps?.onClick) {
-            map.on("click", props.layer.id, (e) => {
-                emit("click", e);
-            });
-        }
-        if (vnodeProps?.onDblclick) {
-            map.on("dblclick", props.layer.id, (e) => {
-                emit("dblclick", e);
-            });
-        }
-        if (vnodeProps?.onMouseenter) {
-            map.on("mouseenter", props.layer.id, (e) => {
-                emit("mouseenter", e);
-            });
-        }
-        if (vnodeProps?.onMouseleave) {
-            map.on("mouseleave", props.layer.id, (e) => {
-                emit("mouseleave", e);
-            });
-        }
-    });
+    }
+
+    addLayer();
 });
 
 onUnmounted(() => {