+++ src/runtime/components/Source.vue
@@ -1,7 +1,7 @@
 <!-- eslint-disable vue/multi-word-component-names -->
 <script setup lang="ts">
-import { AnySourceData, GeoJSONSource, VectorSourceImpl, ImageSource, RasterSource, AnyLayer, CustomLayerInterface } from "mapbox-gl";
-import { inject, onMounted, onUnmounted, watch, useMapbox } from "#imports";
+import { AnySourceData, GeoJSONSource, VectorSourceImpl, ImageSource, RasterSource, Map } from "mapbox-gl";
+import { inject, onUnmounted, watch, useMapbox, computed, useMapboxInstance } from "#imports";
 
 interface Props {
     sourceId: string;
@@ -12,63 +12,62 @@ const props = defineProps<Props>();
 const mapId = inject<string>("MapID");
 if (!mapId) throw "Mapbox Source must be placed inside a Map component";
 
-onMounted(() => {
-    useMapbox(mapId, (map) => {
-        function addLayer() {
+const mapRef = useMapboxInstance(mapId);
+const sourceExists = computed(() => {
+    return !!mapRef.value?.getSource(props.sourceId);
+});
+
+
+useMapbox(mapId, (map: Map) => {
+    function addSource() {
+        if (!sourceExists.value) {
             map?.addSource(props.sourceId, props.source);
         }
-
-        addLayer();
-    });
+    }
+    addSource();
+    map.on('style.load', () => {
+        addSource();
+    })
 });
 
 onUnmounted(() => {
-    useMapbox(mapId, (map) => {
-        map.getStyle().layers
-        .filter((layer) => (layer as Exclude<AnyLayer, CustomLayerInterface>).source === props.sourceId)
-        .forEach((layer) => {
-            map.removeLayer(layer.id);
-        })
-        map.removeSource(props.sourceId);
-    });
+    mapRef.value?.removeSource(props.sourceId);
 });
 
 watch(() => props.source, () => {
-    useMapbox(mapId, (map) => {
-        const source = map?.getSource(props.sourceId);
-        if (!source) return;
-        
-        props.source.type === 'geojson'
-        if (props.source.type === 'geojson' && props.source.data) {
-            // @ts-ignore TODO: Figure out how to type this. Seems like mapbox issue
-            (source as GeoJSONSource).setData(props.source.data);
+    const source = mapRef.value?.getSource(props.sourceId);
+    if (!source) return;
+
+    props.source.type === 'geojson'
+    if (props.source.type === 'geojson' && props.source.data) {
+        // @ts-ignore TODO: Figure out how to type this. Seems like mapbox issue
+        (source as GeoJSONSource).setData(props.source.data);
+    }
+    if (props.source.type === 'vector' && (props.source.tiles || props.source.url)) {
+        if (props.source.tiles) {
+            (source as VectorSourceImpl).setTiles(props.source.tiles);
         }
-        if (props.source.type === 'vector' && (props.source.tiles || props.source.url)) {
-            if (props.source.tiles) {
-                (source as VectorSourceImpl).setTiles(props.source.tiles);
-            }
-            if (props.source.url) {
-                (source as VectorSourceImpl).setUrl(props.source.url);
-            }
-            // @ts-ignore TODO: Not even the reload function?
-            (source as VectorSourceImpl).reload();
+        if (props.source.url) {
+            (source as VectorSourceImpl).setUrl(props.source.url);
         }
-        if (props.source.type === 'image' && props.source.url) {
-            (source as ImageSource).updateImage(props.source);
+        // @ts-ignore TODO: Not even the reload function?
+        (source as VectorSourceImpl).reload();
+    }
+    if (props.source.type === 'image' && props.source.url) {
+        (source as ImageSource).updateImage(props.source);
+    }
+    if (props.source.type === 'raster' && (props.source.url || props.source.tiles)) {
+        if (props.source.url) {
+            // @ts-ignore TODO: Types broken once again
+            (source as RasterSource).setUrl(props.source.url);
         }
-        if (props.source.type === 'raster' && (props.source.url || props.source.tiles)) {
-            if (props.source.url) {
-                // @ts-ignore TODO: Types broken once again
-                (source as RasterSource).setUrl(props.source.url);
-            }
-            if(props.source.tiles) {
-                // @ts-ignore TODO: Types broken once again
-                (source as RasterSource).setTiles(props.source.tiles);
-            }
-            // @ts-ignore TODO: Not even the reload function?
-            (source as RasterSource).reload();
+        if(props.source.tiles) {
+            // @ts-ignore TODO: Types broken once again
+            (source as RasterSource).setTiles(props.source.tiles);
         }
-    })
+        // @ts-ignore TODO: Not even the reload function?
+        (source as RasterSource).reload();
+    }
 })
 </script>