+++ src/runtime/composables/defineMapboxMarker.ts
@@ -1,9 +1,6 @@
-import { ref, Ref } from 'vue';
-import { watch } from 'vue';
-import { useNuxtApp } from '#app';
-import { Marker, MarkerOptions } from "mapbox-gl";
-import { inject } from 'vue';
-import { useMapbox } from './useMapbox';
+import { Ref } from 'vue';
+import { inject, isRef, onUnmounted, ref, useMapbox, useMapboxMarkerRef, useNuxtApp, watch } from '#imports';
+import { MarkerOptions, Marker } from 'mapbox-gl';
 
 
 
@@ -18,59 +15,103 @@ import { useMapbox } from './useMapbox';
  * @param mapID Optionally pass the mapID manually. Will be auto injected if component is nested in MapboxMap
  */
 export function defineMapboxMarker(
-    markerID: string, options: MarkerOptions & { lnglat: mapboxgl.LngLatLike }, markerHTML: Ref<HTMLElement | null>, callback?: (marker: Marker) => void, mapID?: string ): Ref<Marker | null> | undefined
+    markerID: string,
+    options: MarkerOptions & { lnglat: mapboxgl.LngLatLike } | Ref<MarkerOptions & { lnglat: mapboxgl.LngLatLike }>,
+    markerHTML: Ref<HTMLElement | null>,
+    callback?: (marker: Marker) => void,
+    mapID?: string ): Ref<Marker | null> | undefined
 
 /**
  * Create a new Marker instance for a component. Will be automatically added to map if it is nested in MapboxMap
  * 
  * @param mapID Optionally pass the mapID manually. Will be auto injected if component is nested in MapboxMap
  */
-export function defineMapboxMarker(markerID: string, options: MarkerOptions & { lnglat: mapboxgl.LngLatLike }, markerHTML?: undefined, callback?: undefined, mapID?: string ): Marker
+export function defineMapboxMarker(markerID: string,
+    options: MarkerOptions & { lnglat: mapboxgl.LngLatLike } | Ref<MarkerOptions & { lnglat: mapboxgl.LngLatLike }>,
+    markerHTML?: undefined,
+    callback?: undefined, mapID?: string ): Marker
 
-export function defineMapboxMarker(markerID: string, options: MarkerOptions & { lnglat: mapboxgl.LngLatLike }, markerHTML?: Ref<HTMLElement | null> | undefined, callback?: Function, mapID: string = ""): any {
-    const mapId = inject<string>('MapID')
+//TODO: MutationObserver on html so it is reactive
+export function defineMapboxMarker(markerID: string, options: MarkerOptions & { lnglat: mapboxgl.LngLatLike } | Ref<MarkerOptions & { lnglat: mapboxgl.LngLatLike }>, markerHTML?: Ref<HTMLElement | null> | undefined, callback?: Function, mapID: string = ""): any {
     if (!useNuxtApp().$mapboxInitMarker) return; // So we dont run on server.
+    const markerRef = ref<Marker | null>(null);
+    const mapId = inject<string>('MapID')
 
-    const markerRef = ref<Marker | null>(null)
-    if (markerHTML) {
-        if (markerHTML.value) {
-            markerHTML.value.remove()
-
-            useNuxtApp().$mapboxInitMarker(markerID, {element: markerHTML.value, ...options})
-            markerRef.value = useNuxtApp().$mapboxMarkerInstances().value[markerID]
-            if (callback)
-                callback(markerRef.value)
-
-            useMapbox(mapId || mapID, (map) => {
-                if (markerRef.value)
-                    markerRef?.value?.addTo(map)
-            })
-        }
-        watch(markerHTML, () => {
+    function initMarker() {
+        const markerOptions = isRef(options) ? options.value : options;
+        if (markerHTML) {
             if (markerHTML.value) {
                 markerHTML.value.remove()
-
-                useNuxtApp().$mapboxInitMarker(markerID, {element: markerHTML.value, ...options})
+    
+                useNuxtApp().$mapboxInitMarker(markerID, {element: markerHTML.value, ...markerOptions})
                 markerRef.value = useNuxtApp().$mapboxMarkerInstances().value[markerID]
                 if (callback)
                     callback(markerRef.value)
-
+    
                 useMapbox(mapId || mapID, (map) => {
                     if (markerRef.value)
                         markerRef?.value?.addTo(map)
                 })
             }
-        })
-        return markerRef;
-    } else {
-        useNuxtApp().$mapboxInitMarker(markerID, options)
-        const marker = useNuxtApp().$mapboxMarkerInstances().value[markerID]
+            watch(markerHTML, () => {
+                if (markerHTML.value) {
+                    markerHTML.value.remove()
     
-        useMapbox(mapId || mapID, (map) => {
-            marker.setLngLat(options.lnglat);
-            marker.addTo(map);
-        })
-        return marker;
+                    useNuxtApp().$mapboxInitMarker(markerID, {element: markerHTML.value, ...options})
+                    markerRef.value = useNuxtApp().$mapboxMarkerInstances().value[markerID]
+                    if (callback)
+                        callback(markerRef.value)
+    
+                    useMapbox(mapId || mapID, (map) => {
+                        if (markerRef.value)
+                            markerRef?.value?.addTo(map)
+                    })
+                }
+            })
+            return markerRef;
+        } else {
+            useNuxtApp().$mapboxInitMarker(markerID, markerOptions)
+            const marker = useNuxtApp().$mapboxMarkerInstances().value[markerID]
+        
+            useMapbox(mapId || mapID, (map) => {
+                marker.setLngLat(markerOptions.lnglat);
+                marker.addTo(map);
+            })
+            return marker;
+        }
     }
 
+    if (isRef(options)) {
+        // TODO: Watch marker properties and update options ref
+        watch(options, (newOptions, oldOptions) => {
+            const currentMarker = useMapboxMarkerRef(markerID);
+            if (newOptions.draggable !== oldOptions?.draggable && newOptions.draggable !== undefined) {
+                currentMarker.value?.setDraggable(newOptions.draggable);
+            }
+            if (newOptions.offset !== oldOptions?.offset && newOptions.offset !== undefined) {
+                currentMarker.value?.setOffset(newOptions.offset);
+            }
+            if (newOptions.pitchAlignment !== oldOptions?.pitchAlignment && newOptions.pitchAlignment !== undefined) {
+                currentMarker.value?.setPitchAlignment(newOptions.pitchAlignment);
+            }
+            if (newOptions.rotationAlignment !== oldOptions?.rotationAlignment && newOptions.rotationAlignment !== undefined) {
+                currentMarker.value?.setRotationAlignment(newOptions.rotationAlignment);
+            }
+            if (newOptions.rotation !== oldOptions?.rotation && newOptions.rotation !== undefined) {
+                currentMarker.value?.setRotation(newOptions.rotation);
+            }
+            if (newOptions.lnglat !== oldOptions?.lnglat) {
+                currentMarker.value?.setLngLat(newOptions.lnglat);
+            }
+        }, { deep: true });
+    }
+
+    onUnmounted(() => {
+        const currentMarker = useMapboxMarkerRef(markerID);
+        if (currentMarker.value) {
+            currentMarker.value.remove();
+            delete useNuxtApp().$mapboxMarkerInstances().value[markerID];
+        }
+    })
+    return initMarker();
 }