+++ src/runtime/composables/mapboxRef.ts
@@ -1,12 +1,16 @@
 import { Map } from "mapbox-gl";
-import { _useMapboxInstances } from "./useMapboxInstances";
-import { computed, ComputedRef } from "vue"
+import { _useMapboxInstances } from "#imports";
+import { computed, ComputedRef, Ref } from "vue"
 
-export function useMapboxInstance(mapID: string): ComputedRef<Map | null>{
+export function useMapboxRef(mapID: string): Ref<Map | null>{
     const map = computed(() => {
         return _useMapboxInstances()?.value[mapID]?.map || null
     })
     return map
+}
+
+export function useMapboxInstance(mapID: string): Ref<Map | null>{
+    return useMapboxRef(mapID);
 }
 
 export function _useMapboxInstanceWithLoaded(mapID: string): ComputedRef<{ map: Map, loaded: boolean } | null>{